Add example proving knowledge of cubic equation solution
The example of knowing a solution to the cubic equation of `x^3 + x + 5 = y`
is effectively the first thing you come across when dissecting the zkSNARKs
knowledge if your background is in blockchain. This is probably thanks
to Vitalik who decomposes the problem into the primitives in his blog
post [here](https://medium.com/@VitalikButerin/quadratic-arithmetic-programs-from-zero-to-hero-f6d558cea649).

This commit adds this as a working example to Spartan repo.

Signed-off-by: Jakub Konka <kubkon@jakubkonka.com>